+++ docs/swagger.yaml
@@ -110,6 +110,18 @@ info:
   title: NCN Lifecycle Management API
   version: "1.0"
 paths:
+  /v1/liveness:
+    get:
+      consumes:
+      - application/json
+      produces:
+      - application/json
+      responses:
+        "204":
+          description: ""
+      summary: K8s Liveness endpoint
+      tags:
+      - Misc
   /v1/ncns/hooks:
     get:
       consumes:
@@ -234,6 +246,40 @@ paths:
       summary: End to end rolling rebuild ncns
       tags:
       - NCN Lifecycle Events
+  /v1/readiness:
+    get:
+      consumes:
+      - application/json
+      produces:
+      - application/json
+      responses:
+        "204":
+          description: ""
+        "500":
+          description: Internal Server Error
+          schema:
+            $ref: '#/definitions/ResponseError'
+      summary: K8s Readiness endpoint
+      tags:
+      - Misc
+  /v1/version:
+    get:
+      consumes:
+      - application/json
+      produces:
+      - application/json
+      responses:
+        "200":
+          description: OK
+          schema:
+            $ref: '#/definitions/ResponseOk'
+        "500":
+          description: Internal Server Error
+          schema:
+            $ref: '#/definitions/ResponseError'
+      summary: Get version of cray-nls service
+      tags:
+      - Misc
   /v1/workflows:
     get:
       consumes:
