+++ main.go
@@ -24,9 +24,9 @@
 package main
 
 import (
-	"github.com/Cray-HPE/cray-nls/bootstrap"
 	_ "github.com/Cray-HPE/cray-nls/docs"
-	"github.com/Cray-HPE/cray-nls/utils"
+	"github.com/Cray-HPE/cray-nls/src/bootstrap"
+	"github.com/Cray-HPE/cray-nls/src/utils"
 	"github.com/joho/godotenv"
 	"go.uber.org/fx"
 )
