+++ src/api/controllers/v1/misc/misc.go
@@ -26,8 +26,8 @@ package controllers_v1
 import (
 	"net/http"
 
-	"github.com/Cray-HPE/cray-nls/api/services"
-	"github.com/Cray-HPE/cray-nls/utils"
+	"github.com/Cray-HPE/cray-nls/src/api/services"
+	"github.com/Cray-HPE/cray-nls/src/utils"
 	"github.com/gin-gonic/gin"
 )
