+++ src/api/routes/ncn.go
@@ -26,8 +26,8 @@
 package routes
 
 import (
-	controllers_v1 "github.com/Cray-HPE/cray-nls/api/controllers/v1"
-	"github.com/Cray-HPE/cray-nls/utils"
+	controllers_v1 "github.com/Cray-HPE/cray-nls/src/api/controllers/v1"
+	"github.com/Cray-HPE/cray-nls/src/utils"
 )
 
 // NcnRoutes struct
